+++ accepton/client.py
@@ -2,13 +2,15 @@ from .api.planning import Planning
 from .api.promotion import Promotion
 from .api.querying import Querying
 from .api.refunding import Refunding
+from .api.subscribing import Subscribing
 from .api.tokenization import Tokenization
 from .version import VERSION
 
 __all__ = ["Client"]
 
 
-class Client(Planning, Promotion, Querying, Refunding, Tokenization):
+class Client(Planning, Promotion, Querying, Refunding, Subscribing,
+             Tokenization):
     def __init__(self, api_key=None, environment="production"):
         self.api_key = api_key
         self.environment = environment